Extract attribute utilities (and fix copyrightHolder)
