+++ src/app/content/components/Attribution.tsx
@@ -17,6 +17,7 @@ import { findDefaultBookPage, getBookPageUrlAndParams } from '../utils';
 import { contentTextStyle } from './Page/PageContent';
 import { disablePrint } from './utils/disablePrint';
 import { wrapperPadding } from './Wrapper';
+import { bookIdsWithSpecialAttributionText, compensateForUTC, getAuthors, getPublishDate, } from './utils/attributionValues';
 
 const detailsMarginTop = 3;
 const desktopSpacing = 1.8;
@@ -96,44 +97,6 @@ const AttributionDetails = styled(Details)`
 
   ${disablePrint}
 `;
-// date is initialized as UTC, conversion to local time can change the date.
-// this compensates
-function compensateForUTC(date: Date): void {
-  date.setMinutes(date.getMinutes() + date.getTimezoneOffset());
-}
-
-function getPublishDate(book: Book): Date | null {
-  if (!('publish_date' in book)) {
-    return null;
-  }
-  const date = new Date(book.publish_date);
-
-  compensateForUTC(date);
-  return date;
-}
-
-function getAuthors(book: Book) {
-  if (!('authors' in book)) {
-    return [];
-  }
-  const seniorAuthors = book.authors.filter((author) => author.value.senior_author);
-
-  return seniorAuthors.length > 0
-    ? seniorAuthors
-    : book.authors.slice(0, 2);
-}
-
-export function attributionValues(book: Book) {
-  const bookPublishDate = getPublishDate(book);
-  const authorsToDisplay = getAuthors(book);
-
-  return {
-    bookTitle: book.title,
-    copyrightHolder: 'OpenStax',
-    bookPublishDate,
-    authorsToDisplay,
-  };
-}
 
 // tslint:disable-next-line:variable-name
 const AttributionContent = htmlMessage('i18n:attribution:text', Content);
@@ -145,24 +108,6 @@ interface Props {
 
 class Attribution extends Component<Props> {
   public container = React.createRef<HTMLDetailsElement>();
-  private bookIdsWithSpecialAttributionText: {
-    [key: string]: {
-      copyrightHolder?: string,
-      originalMaterialLink?: null | string,
-    }
-  } = {
-    '1b4ee0ce-ee89-44fa-a5e7-a0db9f0c94b1': {
-      copyrightHolder: 'The Michelson 20MM Foundation',
-    },
-    '394a1101-fd8f-4875-84fa-55f15b06ba66': {
-      copyrightHolder: 'Texas Education Agency (TEA)',
-      originalMaterialLink: 'https://www.texasgateway.org/book/tea-statistics',
-    },
-    'cce64fde-f448-43b8-ae88-27705cceb0da': {
-      copyrightHolder: 'Texas Education Agency (TEA)',
-      originalMaterialLink: 'https://www.texasgateway.org/book/tea-physics',
-    },
-  };
   private toggleHandler: undefined | (() => void);
 
   public componentDidMount() {
@@ -231,7 +176,7 @@ class Attribution extends Component<Props> {
       currentPath: currentPageUrl,
       introPageUrl,
       originalMaterialLink: null,
-      ...this.bookIdsWithSpecialAttributionText[book.id] || {},
+      ...bookIdsWithSpecialAttributionText[book.id] || {},
     };
   };
 }
